refactor: 운동 수정 API에서 중첩 record 패턴 적용
