--- src/main/java/umc/cockple/demo/domain/exercise/dto/ExerciseAddrUpdateCommand.java
@@ -1,12 +1,0 @@
-package umc.cockple.demo.domain.exercise.dto;
-
-import lombok.Builder;
-
-@Builder
-public record ExerciseAddrUpdateCommand(
-        String roadAddress,
-        String buildingName,
-        Double latitude,
-        Double longitude
-) {
-}
